+++ src/cv/com/escola/model/dao/UsuarioDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UsuarioDAO extends CrudDAO<Usuario, Integer>{
 
     void createUserAdminAndUserType();
-    boolean isUsuario(int id, String nome);
+    boolean isUsuario(Long id, String nome);
     int totalTipoUsuario();
     int totalUsuario();
     List<TipoUsuario> usuariosTipo();
